Refactor read time calculation into a separate component; update usage in ContentEditPage and Visualizer
